Working container and system viewer. deletes all  completed. some major gui changes too. This all allows for smooth transitions between forms. You can create new systems/containers in the system/container viewers now.

+++ Container File Optimizer/ContainerViewer.cs
@@ -119,8 +119,6 @@ namespace Container_File_Optimizer
                     //exacute the delete
                     command.ExecuteNonQuery();
 
-                    //message to show the system was deleted
-                    MessageBox.Show("Deleted...");
 
 
                     //close connection
@@ -249,9 +247,6 @@ namespace Container_File_Optimizer
                         //exacute the delete
                         command.ExecuteNonQuery();
 
-                        //message to show the system was deleted
-                        MessageBox.Show("Deleted file");
-
 
                         //close connection
                         connection.Close();
@@ -362,37 +357,66 @@ namespace Container_File_Optimizer
         {
             if (listBoxFiles.SelectedItems.Count > 0)
             {
+                MessageBox.Show("Deleted file: " + listBoxFiles.SelectedItem.ToString() + " from " + listBoxContainerViewer.SelectedItem.ToString());
                 deleteFile(fileIDCollection[listBoxFiles.SelectedIndex], appIDCollection[listBoxContainerViewer.SelectedIndex]);
+                //Clear the list box and id collection dictionary 
+                listBoxFiles.Items.Clear();
+                fileIDCollection.Clear();
+                // To populate the list box
+                GetFileNames(appIDCollection[listBoxContainerViewer.SelectedIndex]);
+                // to populate the dictionary
+                GetFileIDS(appIDCollection[listBoxContainerViewer.SelectedIndex]);
             }
-            //Clear the list box and id collection dictionary 
-            listBoxFiles.Items.Clear();
-            fileIDCollection.Clear();
-            // To populate the list box
-            GetFileNames(appIDCollection[listBoxContainerViewer.SelectedIndex]);
-            // to populate the dictionary
-            GetFileIDS(appIDCollection[listBoxContainerViewer.SelectedIndex]);
+            else
+            {
+                MessageBox.Show("Error: No file was selected to delete. Please first select a file.");
+            }
+
         }
 
         private void buttonDeleteContainer_Click(object sender, EventArgs e)
         {
-            deleteApplication(appIDCollection[listBoxContainerViewer.SelectedIndex]);
-            for (int i = 0; i < listBoxContainerViewer.Items.Count; i++)
+            if(listBoxContainerViewer.Items.Count > 0)
             {
+                MessageBox.Show("Deleted container: " + listBoxContainerViewer.SelectedItem.ToString());
+                // First delete all the files in that app if they only occur once which is done through the deleteFiles() function
+                for (int i = 0; i < listBoxFiles.Items.Count; i ++)
+                {
+                    deleteFile(fileIDCollection[i], appIDCollection[listBoxContainerViewer.SelectedIndex]);
+                }
+                // Next delete the empty container
+                deleteApplication(appIDCollection[listBoxContainerViewer.SelectedIndex]);
 
-                deleteFile(fileIDCollection[i], appIDCollection[listBoxContainerViewer.SelectedIndex]);
+
+                // Clear form objects and collections
+
+                appIDCollection.Clear();
+                listBoxContainerViewer.Items.Clear();
+
+                fileIDCollection.Clear();
+                listBoxFiles.Items.Clear();
+
+
+
+
+
+
+                // Populate list box and app ids collection dictionary
+                ViewContainers();
+
+            }
+            else
+            {
+                MessageBox.Show("Error: No container was selected to be deleted...");
             }
 
+        }
 
-
-            listBoxFiles.Items.Clear();
-            fileIDCollection.Clear();
-            listBoxContainerViewer.Items.Clear();
-
-            // To populate the list box
-            GetFileNames(appIDCollection[listBoxContainerViewer.SelectedIndex]);
-            // to populate the dictionary
-            GetFileIDS(appIDCollection[listBoxContainerViewer.SelectedIndex]);
-            ViewContainers();
+        private void buttonNewContainer_Click(object sender, EventArgs e)
+        {
+            this.Close();
+            NewContainer newContainerForm = new NewContainer();
+            newContainerForm.Show();
         }
     }
 }
